requirements.txt
fastapi
pyjwt
uvicorn[standard]
bcrypt
sqlalchemy